maison/requirements.txt
fastapi==0.95.2
uvicorn>=0.23,<0.24
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==1.10.11
httpx==0.24.1
jinja2
aiofiles
python-multipart
prometheus-fastapi-instrumentator
pydantic[email]
